fix(ci): resolve mypy errors in chat service/knowledge/qwen
Type generator return for heartbeat runner, restore duck-typed getattr on
knowledge proposals, and tighten queue/item typing so backend CI passes.

+++ src/adaptive_rag/chat/qwen.py
@@ -1080,7 +1080,7 @@ def _retry_backoff_seconds(
             except ValueError:
                 pass
     # attempt 0 → 0.5s, 1 → 1s, 2 → 2s (capped)
-    return min(0.5 * (2**attempt), 8.0)
+    return float(min(0.5 * (2**attempt), 8.0))
 
 
 def _elapsed_ms(started: float) -> int:
